+++ rts/c/multicore_util.h
@@ -36,7 +36,7 @@ static inline int getrusage_thread(struct rusage *rusage)
 }
 
 /* returns the number of logical cores */
-static const int num_processors()
+static int num_processors()
 {
 #if  defined(_WIN32)
 /* https://docs.microsoft.com/en-us/windows/win32/api/sysinfoapi/ns-sysinfoapi-system_info */